Make the name of "Korean Rice Cakes" more descriptive
As I was going through items on Poppypedia, I came across "Korean Rice Cakes." I found it kind of weird that a lot of other food (like mochi) had its name in its language of origin, but not tteok. From the recipe and art, this is most similar to garaetteok, which is used in tteokbokki and tteokguk. I've used "garaetteok" here, but tteok is the general word for rice cake.